% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{acre_to_hectare}
\alias{acre_to_hectare}
\title{Convert area from Acres to hectares}
\usage{
acre_to_hectare(acres)
}
\arguments{
\item{acres}{a value or numeric vector of area in acres}
}
\value{
a value or numeric vector of the same area in hectares
}
\description{
Convert area from Acres to hectares
}
