% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{bearing}
\alias{bearing}
\title{Bearing angle
Calculate the angle between two geographic points relative to true north (clockwise)}
\usage{
bearing(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of starting point in decimal degrees.}

\item{lon1}{Longitude of starting point in decimal degrees}

\item{lat2}{Latitude of end point in decimal degrees}

\item{lon2}{Longitude of end point in decimal degrees.}
}
\value{
A single number in degrees, representing the direction from start to end point. 
0 degrees means the end point is due north of the start point. 90 degrees means it is due east.
}
\description{
Bearing angle
Calculate the angle between two geographic points relative to true north (clockwise)
}
\examples{
p1 <- c("lon"= -80.411367,"lat" = 43.640741)
p2 <- c("lon" = -80.410466 , "lat" =  43.641394)
mybearing <- bearing(
  lat1 = p1["lat"], 
  lon1 = p1["lon"], 
  lat2 = p2["lat"], 
  lon2 = p2["lon"])

}
