% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_categories}
\alias{check_categories}
\title{Check category dictionary}
\usage{
check_categories(
  table_name = "site_info",
  inputdf = NULL,
  mytoken = getOption("drivesR.default.directustoken"),
  ...
)
}
\arguments{
\item{table_name}{The table (collection) identifier}

\item{inputdf}{A data frame of prospective rows to be added to the table. 
If NULL, the function checks the category and column dictionaries.}

\item{mytoken}{API token, formatted as "Bearer apitoken"}

\item{...}{Arguments passed to api_request.}
}
\value{
If inputdf = NULL, it returns a data frame describing conditions and outcomes corresponding to 
particular tables and fields. 
If inputdf is provided, it returns a list with a dataframe of conditions (checkdf) and a list of 
categorical values in the input data that are missing from the category dictionary. This list can
be used to ammend the category dictionary.
}
\description{
Checks for discrepancies between the category dictionary and 
column dictionary and input data.
}
\examples{
#Not run: check_categories("site_info") 

}
