% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_column_names}
\alias{check_column_names}
\title{Check column names}
\usage{
check_column_names(
  table_name = "site_info",
  inputdf = NULL,
  mytoken = "Bearer apitoken"
)
}
\arguments{
\item{table_name}{Table identifier, such as "site_info"}

\item{inputdf}{Dataframe to be evaluated.}

\item{mytoken}{directus API token, as "Bearer apitoken"}

\item{...}{Arguments passed to get_db_info.}
}
\value{
A dataframe of checks, outcomes, and relevant fields.
}
\description{
Quality control step to check column names in a data frame to be imported into Directus.
}
\examples{
#Not run: check_column_names("site_info", mydf)
}
