% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_dictionary}
\alias{check_dictionary}
\title{Check dictionary metadata}
\usage{
check_dictionary(table_name = "site_info", mytoken = "Bearer {myAPItoken}")
}
\arguments{
\item{table_name}{Table (collection) name.}

\item{mytoken}{Directus token. Can be set in the function or via set_default_token()E}
}
\value{
A data frame with a set of conditions, outcome (TRUE = pass or FALSE = fails), and 
fields that fail the condition. In case of mismatched field names, there are separate 
columns for fields in the column_dictionary (dcols) and Directus table schema (tcols).
}
\description{
Check for discrepancies between collection metadata stored in Directus
 and in data dictionaries.
}
