% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_dups}
\alias{check_dups}
\title{Subset a dataframe for rows that are duplicated across a set of columns.}
\usage{
check_dups(mydf, checkcols = NULL, dupvec = NULL, dupTFmatchesInput = TRUE)
}
\arguments{
\item{mydf}{A dataframe.}

\item{checkcols}{A vector of column names that you want to check for duplicates. Will also work with column indices. Alternative to dupvec.}

\item{dupvec}{A vector to use to check for duplicate values. Length(dupvec) = nrow(mydf). Alternative to column names.}

\item{dupTFmatchesInput}{If TRUE, the TRUE/FALSE vector indicating duplicated values corresponds 
to the input dataframe. If FALSE, it corresponds to the subsetted dataframe containing duplicate values. 
values}
}
\value{
Returns a list containing the number of unique column combinations that have duplicate values (ndups),
a subsetted dataframe containing all duplicated values,
and a TRUE FALSE vector to facilitate removal of duplicate rows (created by duplicated). 
This vector can correspond to the input dataframe (mydf) or the output dataframe (dupdf),
depending on the choice for dupTFmatchesInput.
}
\description{
Subset a dataframe for rows that are duplicated across a set of columns.
}
\examples{
testdf <- data.frame("col1" = c("spam","eggs","spam","spam"),"col2" = c(1,2,1,1),"col3" = c("cat","fish","dog","cat"),"col4" = c(1,2,3,4))
check_dups(testdf, checkcols = c("col1","col2"))
check_dups(testdf, checkcols = c(1,2,3))
}
