% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_fk_values}
\alias{check_fk_values}
\title{Check foreign key values for a staged table.}
\usage{
check_fk_values(
  table_name = NULL,
  inputdf = NULL,
  mytoken = getOption("drivesR.default.directustoken"),
  myurl = getOption("drivesR.default.url")
)
}
\arguments{
\item{table_name}{Name of the table in the DRIVES database.}

\item{inputdf}{Data frame containing foreign key colulmns.}

\item{mytoken}{Directus token.}

\item{myurl}{Directus base url.}
}
\value{
A list of columns with foreign key constraints and 
a dataframe of subsetted rows from inputdf that violate
foreign key constraints--or NULL if all rows pass. 
The function throws out messages about the results of each foreign
key columns.
}
\description{
Check foreign key values for a staged table.
}
\examples{
# not run: fkcheck <- check_fk_values(
#"crop_variety_info", inputdf = staged_df)
}
