% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_table_contents}
\alias{check_table_contents}
\title{Check table contents before import}
\usage{
check_table_contents(
  table_name = "site_info",
  inputdf = NULL,
  mytoken = "Bearer {myAPItoken}"
)
}
\arguments{
\item{table_name}{Name of the table in Directus}

\item{inputdf}{data frame intended for the Directus table}

\item{mytoken}{User-specific Directus token. Can set with set_default_token()}
}
\value{
A list with columns and rows failing each constraint. 
1) is_nullable
2) max_length 
3) is_unique
4) data type.
}
\description{
This function checks contents of a table (usually imported as a CSV) for
possible failures in validation constraints. 
The constraints checked are:
1) is_nullable
2) max_length (if applicable)
3) is_unique
4) data type.
Column names and foreign key values are checked with a different function.
}
\examples{
#not run: mydf <- read.csv("site_info_to_add.csv")
#not run: check_table_contents("site_info", inputdf = mydf)
}
