% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_treatment_years}
\alias{check_treatment_years}
\title{Check for mistakes in treatment components and years.}
\usage{
check_treatment_years(db = NULL)
}
\arguments{
\item{db}{A list containing named data frames treatment_id_info, treatment_id_components,
and experimental_unit_treatments. Because this function is intended to check for mistakes before
uploading to the database, this list must be supplied by the user.}
}
\value{
A list with subsetted problem rows of experimental_unit_treatments, treatment_id_components, and the
combined dataframe with treatment components by experimental unit.
}
\description{
Harmonization steps require that:
1) Expanding by start and end years in experimental_unit_treatments
results in a single row per treatment_id per unit_id per year (one for treatmentID1, one for treatmentID2)
2) Expanding by start and end years in treatment_id_components results
in 1 level per treatment type and treatment id per year.
3) When these expanded tables are combined, there is only 1 level per treatment type, unit id, treatment id, and year.
}
\examples{
# not run trt_check <- check_treatment_years(db)
}
