% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{date_formats}
\alias{date_formats}
\title{Date formats}
\format{
A list of vectors
\describe{
  \item{mdy}{Vector of month-day-year formats}
  \item{yyyymd}{Vector of year-month-day formats}
  \item{bmo}{Vector with letter or word months}
  \item{dmy}{Vector of day-month-year formats}
  \item{unlikely}{Vector of unlikely date formats}
 
  ...
}
}
\usage{
date_formats
}
\description{
A list of possible date formats.
}
\keyword{datasets}
