% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_edit_functions.R
\name{delete_rows}
\alias{delete_rows}
\title{Delete rows in a table based on their primary key value}
\usage{
delete_rows(
  table_name = NULL,
  pkvec = NULL,
  check_only = FALSE,
  mytoken = getOption("drivesR.default.directustoken"),
  ...
)
}
\arguments{
\item{table_name}{The table identifier in directus}

\item{pkvec}{A vector of primary keys for rows you want to delete in the table.}

\item{check_only}{TRUE or FALSE indicating whether you want to inspect the rows instead of deleting them.}

\item{mytoken}{Directus api token, formatted as "Bearer myapitoken"}

\item{...}{Arguments passed to get_db_table and api_request.}
}
\value{
If check_only = TRUE, returns a data frame of the table subset corresponding to the primary 
keys in pkvec. if check_only = FALSE, the function performs delete operations on primary keys
within the specified table.
}
\description{
Delete rows in a table based on their primary key value
}
\examples{
# not run: fakerowjson <- make_insert_json(data.frame("site_id" = c("fake_id1","fake_id2")))
# not run: api_request("POST","items/site_info", fakerowjson)
# not run: testrun <- delete_rows("site_info",c("fake_id1","fake_id2"), check_only = TRUE)
# not run: delete_rows("site_info",c("fake_id1","fake_id2"), check_only = FALSE)
}
