% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivesR-package.R
\docType{package}
\name{drivesR-package}
\alias{drivesR}
\alias{drivesR-package}
\title{drivesR: DRIVES Database Workflow}
\description{
Contains various functions to access and harmonize data from the DRIVES database. Also contains functions used to prepare data, perform quality checks, and manage the database structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://drives-project.github.io/drivesR/}
}

}
\author{
\strong{Maintainer}: Katherine Muller \email{katherine.muller@usda.gov} (\href{https://orcid.org/0000-0001-9144-3867}{ORCID})

}
\keyword{internal}
