% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{expand_years}
\alias{expand_years}
\title{Expand data frame by start and end years}
\usage{
expand_years(
  mydf = NULL,
  na_end_year = 2021,
  sycol = "start_year",
  eycol = "end_year",
  ycol = "year"
)
}
\arguments{
\item{mydf}{A data frame containing start- and end-year columns}

\item{na_end_year}{A year to use when end-year is NA. 
Start years must not have any NAs.}

\item{sycol}{Name of the start-year column. Default is start_year.}

\item{eycol}{Name of the end-year column. Default is end_year.}

\item{ycol}{Name of the new column for individual years. Default is year.}
}
\value{
A data-frame with separate rows for each year between start and end year. 
Start and end year columns are removed.
}
\description{
Converts a table with time ranges specified as start- and end-year column into 
a longer table with separate rows for each year.
}
\examples{
#Not run: exp_unit_trt <- get_db_table("experimental_units_treatments") 
#Not run: exp_trt_yearly <- expand_years(exp_unit_trt)
}
