% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius conversion}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Farenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C(c(50,63,23))
}
