% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{ft_to_m}
\alias{ft_to_m}
\title{Convert length from feet to meters}
\usage{
ft_to_m(ft_len)
}
\arguments{
\item{ft_len}{Length in feet}
}
\value{
Length in meters
}
\description{
Convert length from feet to meters
}
\examples{
ft_to_m(8);
ft_to_m(c(8,9,10))
}
