% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_column_dict_for_table}
\alias{get_column_dict_for_table}
\title{Get column dictionary for a specified table
As an alternative to subsetting.}
\usage{
get_column_dict_for_table(
  table_name = "site_info",
  myurl = getOption("drivesR.default.url")
)
}
\arguments{
\item{table_name}{}

\item{myurl}{Base url for the Directus database.}
}
\value{
A dataframe consisting of the subsetted column dictionary for the corresponding table.
}
\description{
Get column dictionary for a specified table
As an alternative to subsetting.
}
\examples{
#not run: cdict <- get_column_dict_for_table("site_info")

}
