% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_table_from_req}
\alias{get_table_from_req}
\title{Download content from an API fetch request into a data frame}
\usage{
get_table_from_req(apirequest = NULL)
}
\arguments{
\item{apirequest}{An API request made with httr or api_request() with GET or SEARCH.}
}
\value{
A data-frame with contents from the API request.
}
\description{
This is useful for api_request() that involves a json_body object, such as filtering for conditions in a table.
}
