% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{harmonize_treatments}
\alias{harmonize_treatments}
\title{Harmonize treatment components}
\usage{
harmonize_treatments(db = NULL)
}
\arguments{
\item{db}{A list of database tables containing named dataframes treatment_id_info and treatment_id_components.
If left NULL, these tables are imported from Directus.}
}
\value{
A data frame with columns for site, year, treatmentID1, and treatmentID2 and all treatment types
described in the table site_treatment_type_info
}
\description{
Combines information stored in the tables treatment_id_info and treatment_id_components into a time-series dataset
describing combined experimental components for every year.
}
\examples{
#not run: tcw <- harmonize_treatments()
}
