% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{harmonize_treatments_units}
\alias{harmonize_treatments_units}
\title{Harmonize treatment and unit information}
\usage{
harmonize_treatments_units(
  db = NULL,
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{db}{A list of database tables containing named dataframes treatment_id_info,
treatment_id_components, and experimental_unit_treatments.
If left NULL, these tables are imported from Directus.}

\item{mytoken}{Directus token, formatted as "Bearer myapitoken". Can be set with set_default_token()}
}
\value{
A data frame with one row per unit ID per year and all treatment components in columns.
}
\description{
Combines information from database tables to describe which treatment components
were applied to which experimental units in which years.
}
\examples{
# not run: trt_units <- harmonize_treatment_units()
}
