% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{harmonize_weather}
\alias{harmonize_weather}
\title{Harmonize daily weather data}
\usage{
harmonize_weather(weather_daily = NULL)
}
\arguments{
\item{weather_daily}{A dataframe corresponding to the weather_daily table
in the DRIVES database. If NULL, this table is downloaded from Directus.}
}
\value{
A data frame of mildly processed weather data with weather variables in separate columns
(instead of in separate rows as in the database). The processed table also excludes
weather_station_id, flag, and uid columns from the original database table.
}
\description{
Harmonize daily weather data
}
