% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{harmonize_yields_treatments}
\alias{harmonize_yields_treatments}
\title{Harmonize yield data with experimental treatment data.}
\usage{
harmonize_yields_treatments(
  db = NULL,
  crop_fractions_as_columns = FALSE,
  public = getOption("drivesR.default.public"),
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{db}{#' A list of database tables containing named dataframes treatment_id_info,
 treatment_id_components, experimental_unit_treatments, and crop_yields.
If left NULL, these tables are imported from Directus.}

\item{crop_fractions_as_columns}{#' If TRUE, multiple fractions from the same crop are organized in separate 
columns. If FALSE, multiple fractions from the same crop are organized in separate rows, 
as in the database table.}

\item{public}{TRUE if data are to be downloaded from the publicly available part of the DRIVES 
database. FALSE otherwise. Can be set directly or though options.}

\item{mytoken}{Directus token, formatted as "Bearer myapitoken". Set by set_default_token().}
}
\value{
A dataframe of yield data combined with experimental treatment data, with each treatment
type in a separate column. 
If crop_fractions_as_columns is set to TRUE, 
data describing fractions from the same crop will be described in separate columns, with 
suffix _1 for the primary fraction and _2, _3, etc. for other fractions.
}
\description{
Harmonize yield data with experimental treatment data.
}
\examples{
 # not run: longyield <- harmonize_yields_treatments(crop_fractions_as_columns = FALSE)
# not run: wideyield <- harmonize_yields(crop_fractions_as_columns = TRUE)
}
