% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_utility_functions.R
\name{httr_dry_run}
\alias{httr_dry_run}
\title{API dry run}
\usage{
httr_dry_run(r)
}
\arguments{
\item{r}{An API request through the httr package. More convenient with piping.}
}
\value{
A "request" object, containing a list of fields pertaining to the HTTP request.
}
\description{
Tests an API request without sending it. Useful for troubleshooting.
}
