% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_wrangling.R
\name{identify_date_format}
\alias{identify_date_format}
\title{Identify date formats}
\usage{
identify_date_format(date)
}
\arguments{
\item{date}{Single string containing a date.
For a vector of dates, use lapply or sapply.}
}
\value{
a vector of date formats that match the input date (including MDY and DMY). 
If the date does not match any in date_formats, returns "Unknown Format". This also applies if date is missing any component of day, month, and year.
}
\description{
Identify date formats
}
