% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{import_dictionary_tables}
\alias{import_dictionary_tables}
\title{Import dictionary tables}
\usage{
import_dictionary_tables(
  public = FALSE,
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{public}{True or false indicating whether tables are in the set approved for public access.}
}
\value{
A list of three dictionary tables. One describing database tables, one 
describing columns within tables, and one describing categories within columns.
}
\description{
Import dictionary tables
}
\examples{
# not run: dict <- import_dictionary_tables()
}
