% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{kg_to_lb}
\alias{kg_to_lb}
\title{Convert weight from kg to lb.}
\usage{
kg_to_lb(kg_wt)
}
\arguments{
\item{kg_wt}{Weight in kg.}
}
\value{
Weight in lb.
}
\description{
Convert weight from kg to lb.
}
\examples{
kg_to_lb(1)
kg_to_lb(c(1,2,3))
}
