% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{lb_to_kg}
\alias{lb_to_kg}
\title{Convert weight from pounds to kg}
\usage{
lb_to_kg(lb_wt)
}
\arguments{
\item{lb_wt}{Weight in pounds.}
}
\value{
Weight in kg.
}
\description{
Convert weight from pounds to kg
}
\examples{
lb_to_kg(1);
lb_to_kg(c(1,2,3))
}
