% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{m_to_ft}
\alias{m_to_ft}
\title{Convert length from meters to feet}
\usage{
m_to_ft(m_len)
}
\arguments{
\item{m_len}{Length in meters}
}
\value{
Length in feet.
}
\description{
Convert length from meters to feet
}
\examples{
m_to_ft(5);
m_to_ft(c(5,6,7))
}
