% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{make_collection_json}
\alias{make_collection_json}
\title{Make a collection schema json object from data dictionary tables}
\usage{
make_collection_json(
  columndf = test_column_dict[which(test_column_dict$table_name == "test_cat_info"), ],
  tablerow = test_table_dict[1, ]
)
}
\arguments{
\item{columndf}{A dataframe of schema information about each column (field) from the column dictionary.}

\item{tablerow}{A one-row dataframe of schema information about the table (collection) from the table dictionary.}
}
\value{
A character string for the corresponding data type recognized by the Directus API.
}
\description{
Make a collection schema json object from data dictionary tables
}
\examples{
my_collection_json <- make_collection_json(columndf = test_column_dict[which(test_column_dict$table_name=="test_cat_info"),],
                                            test_table_dict[1,])


}
