% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{make_row_insert_json}
\alias{make_row_insert_json}
\title{Convert a dataframe or tibble to a json object suitable to insert rows (items)
 
Basically just the jsonlite::toJSON function with options set to non-defaults 
for conventience}
\usage{
make_row_insert_json(mydf)
}
\arguments{
\item{mydf}{A dataframe or tibble with quality control steps complete}
}
\value{
A prettified json object
}
\description{
Convert a dataframe or tibble to a json object suitable to insert rows (items)
 
Basically just the jsonlite::toJSON function with options set to non-defaults 
for conventience
}
\examples{
testdf <- data.frame(x = 1:5, y = LETTERS[1:5])
insert_json <- make_row_insert_json(testdf)
# example with API request:
test_insert <- make_row_insert_json(test_cat_info)
# insert_req <- api_request("POST","items/test_cat_info",test_insert)
}
