% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template_df.R
\name{make_template_df}
\alias{make_template_df}
\title{Make a template data frame from a vector of column names}
\usage{
make_template_df(cnames = NULL, numrows = 1)
}
\arguments{
\item{cnames}{A vector of column names}

\item{numrows}{Number of rows the template should have.}
}
\value{
An empty data frame with specified column names
and number of rows.
}
\description{
Make a template data frame from a vector of column names
}
\examples{
cnames <- c("a","b","c")
templatedf <- make_template_df(cnames, 5)
print(templatedf) 
}
