% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{order_rows_by_internal_fk}
\alias{order_rows_by_internal_fk}
\title{Order rows based on an internal foreign key
For use in uploading rows to database tables. This
specifies a row order for uploading that avoids violating
internal foreign key constraints within a table.}
\usage{
order_rows_by_internal_fk(inputdf = NULL, idcol = NULL, fkcol = NULL)
}
\arguments{
\item{inputdf}{A dataframe of rows to be added to a table with an 
internal foreign key constraint.}

\item{idcol}{The primary key column name (e.g., crop_id).}

\item{fkcol}{The internal foreign key column name (e.g., parent_crop_id)}
}
\value{
A vector of row indices for the inputdf, indicating
the order they should be added in.
}
\description{
Order rows based on an internal foreign key
For use in uploading rows to database tables. This
specifies a row order for uploading that avoids violating
internal foreign key constraints within a table.
}
\examples{
#not run: roworder <- order_rows_by_internal_fk(
#inputdf = crop_info,idcol = "crop_id",fkcol = "parent_crop_id")
# not run: walk(roworder,~post_rows("crop_info",inputdf[.x,] ))
}
