% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{order_tables_by_fk}
\alias{order_tables_by_fk}
\title{Order tables by foreign key dependencies
This is used for posting items to collections with 
foreign key constraints. This is to help determine
the order of items to post to avoid foreign key violations.
Tables with internal foreign keys should use order_rows_by_internal_fk}
\usage{
order_tables_by_fk(column_dictionary = NULL)
}
\arguments{
\item{column_dictionary}{A subset of the column_dictionary describing tables to be ordered.}
}
\value{
A data frame with table names in an order that is compatible with between-table
foreign key constraints, along with a TRUE/FALSE column indicating which tables
have internal foreign keys.
}
\description{
Order tables by foreign key dependencies
This is used for posting items to collections with 
foreign key constraints. This is to help determine
the order of items to post to avoid foreign key violations.
Tables with internal foreign keys should use order_rows_by_internal_fk
}
