% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_functions.R
\name{originFromLonLat}
\alias{originFromLonLat}
\title{Get origin XY coordinates from longitude and latitude points.
Used to estimate georeferenced plot geometry from map dimensions and 
points selected from Google Earth imagery.}
\usage{
originFromLonLat(
  lonlatvec,
  lonlatcrs = 4326,
  outcrs = "ESRI:102008",
  outputtype = c("coordinates", "sf")[1]
)
}
\arguments{
\item{lonlatvec}{A numeric vector of length 2 containing longitude followed by latitude 
in decimal degrees. Usually a point taken from Google Earth imagery.}

\item{lonlatcrs}{The Coordinate Reference System for the longitude latitude points. 
The default is WGS 84, ID EPSG 4326.}

\item{outcrs}{The Coordinate Reference System for the output coordinates.
Usually this is a projection to meters. The default is ESRI:102008.}

\item{outputtype}{An option for whether the output should be a 1-row matrix of 
X Y coordinates (default) or an sf dataframe.}
}
\value{
A 1-row matrix or sf object with the input lon/lat points projected to 
a specified CRS. Used as the origin for georeferencing.
}
\description{
Get origin XY coordinates from longitude and latitude points.
Used to estimate georeferenced plot geometry from map dimensions and 
points selected from Google Earth imagery.
}
\examples{
bottomright <- c("lon" = -98.846982 ,"lat" =  19.532234)
myorigin <- originFromLonLat(bottomright)
print(myorigin) 
}
