% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{pg_to_directus_type}
\alias{pg_to_directus_type}
\title{Map data types from Postgresql to Directus schema}
\usage{
pg_to_directus_type(pgtype)
}
\arguments{
\item{pgtype}{A postgres data type as specified in the column dictionary}
}
\value{
A character string with the corresponding datatype recognized by the Directus API.
}
\description{
Map data types from Postgresql to Directus schema
}
\examples{
pg_to_directus_type("array")
}
