% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{post_rows}
\alias{post_rows}
\title{Add rows to a database table}
\usage{
post_rows(table_name = NULL, inputdf = NULL)
}
\arguments{
\item{table_name}{The name of the database table.}

\item{inputdf}{The dataframe containing rows to be added. Contents should have been checked.}
}
\value{
Converts the input df to a json and runs an API post request on the specified table. 
The status code of that request is returned as a message.
}
\description{
This adds all rows in an input dataframe in a single API request. To add
more than 100 rows, use post_rows_in_batches. For use in an interactive R session.
}
