% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{post_rows_in_batches}
\alias{post_rows_in_batches}
\title{Post rows in batches}
\usage{
post_rows_in_batches(
  table_name = "crop_yields",
  inputdf = NULL,
  batchsize = 1000
)
}
\arguments{
\item{table_name}{Table identifier in the database.}

\item{inputdf}{Data frame of rows to be added. This should have passed quality control checks.}

\item{batchsize}{Number of rows to be added at a time. Suggested amount is 1000 rows.
Progress is printed to the console.}
}
\value{
Silent.
}
\description{
To get around issues of posting limits.
}
\examples{
# not run: post_rows_in_batches("my_table", bigdf, 1000)

}
