% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_manipulation.R
\name{print_stringvec_as_commasep}
\alias{print_stringvec_as_commasep}
\title{Print a character vector as a list of quoted strings separated by commas.}
\usage{
print_stringvec_as_commasep(mystringvec)
}
\arguments{
\item{mystringvec}{A character vector.}
}
\value{
The character vector printed in the console in a way that can be copied into a vector in the code.
}
\description{
Usually this is something I would use to trim excess names off a data frame.
}
\examples{
df1 <- as.data.frame(matrix(1:26, nrow = 1, ncol = 26, dimnames = list(NULL,LETTERS)));
print_stringvec_as_commasep(LETTERS);
# Copy the output into the script, remove a handful of values I don't want
# for whatever arbitrary, difficult-to-code reason.
colsIWant <- c("A" ,"B" ,"C" ,"D" ,"M"  ,"S" ,"T" ,"U" ,"V" ,"W" ,"X" ,"Y" ,"Z");
df2 <- df1[,colsIWant]

}
