% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_utility_functions.R
\name{read_directus_pat}
\alias{read_directus_pat}
\title{Import Directus PAT
This imports a Directus Personal Access Token from a file created by generate_directus_token.
It also tests whether the token is valid.}
\usage{
read_directus_pat(filepath = "directus_PAT.txt")
}
\arguments{
\item{filepath}{A path to the file created by generate_directus_token().}
}
\value{
returns a character string that can be used in the function set_default_token().
}
\description{
Import Directus PAT
This imports a Directus Personal Access Token from a file created by generate_directus_token.
It also tests whether the token is valid.
}
\examples{
# not run:
# ## run in command line or read password from a non-indexed location:
# generate_directus_token(savedir = "afolder", savename = "directus.PAT.txt")
# mytoken <- read_directus_pat(file.path("afolder","directus.PAT.txt"))
# set_default_token(mytoken)

}
