% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_utility_functions.R
\name{set_public_access}
\alias{set_public_access}
\title{Set default public access settings
This can be used to change the default public access settings
from what is set up during package loading. 
For the publicly-released package, the default will be
public access.}
\usage{
set_public_access(public = TRUE)
}
\arguments{
\item{public}{TRUE for public access, FALSE for private access.}
}
\value{
Changes the option for "drivesR.default.directustoken"
that is used by various functions.
}
\description{
Set default public access settings
This can be used to change the default public access settings
from what is set up during package loading. 
For the publicly-released package, the default will be
public access.
}
