% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{test_api_token}
\alias{test_api_token}
\title{Verify Directus API token
Used within other functions to check for errors.}
\usage{
test_api_token(
  mytoken = getOption("drivesR.default.directustoken"),
  myurl = getOption("drivesR.default.url"),
  silent = TRUE
)
}
\arguments{
\item{mytoken}{Directus API token, formatted as "Bearer APItoken."}

\item{myurl}{Directus database url. Set with defaults as https://data.drives-network.org}

\item{silent}{Indicates whether messages should be printed. Default FALSE.}
}
\value{
TRUE or FALSE indicating whether mytoken  
produces a successful api request.
}
\description{
Verify Directus API token
Used within other functions to check for errors.
}
\examples{
test_api_token(mytoken = "notavalidtoken", silent=FALSE)
}
