% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_category_dict}
\alias{test_category_dict}
\title{Sample category dictionary}
\format{
A data frame with 3 rows and 4 variables:
\describe{
  \item{table_name}{Table identifier}
  \item{column_name}{Column identifier within table}
  \item{category_level}{Options for the category}
  \item{description}{Text}

  ...
}
}
\usage{
test_category_dict
}
\description{
A data frame with database schema information representative of DRIVES tables.
}
\keyword{datasets}
