% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_column_dict}
\alias{test_column_dict}
\title{Sample column dictionary}
\format{
A data frame with 8 rows and 14 variables:
\describe{
  \item{column_id}{Unique column identifier}
  \item{table_name}{Table identifier}
  \item{column_name}{Name of column within the table}
  \item{column_order}{Integer}
  \item{description}{Text}
  \item{postgres_data_type}{Text category}
  \item{primary_key}{TRUE or FALSE}
  \item{record_identifier}{TRUE or FALSE}
  \item{foreign_key_table}{Table name or NA}
  \item{foreign_key_column}{Column name or NA}
  \item{nullable}{TRUE or FALSE}
  \item{unique_value}{TRUE or FALSE}
  \item{is_category}{TRUE or FALSE}
  \item{auto_increment}{TRUE or FALSE}

  ...
}
}
\usage{
test_column_dict
}
\description{
A data frame with database schema information representative of DRIVES tables.
}
\keyword{datasets}
