% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_table_dict}
\alias{test_table_dict}
\title{Sample table dictionary}
\format{
A data frame with 2 rows and 4 variables:
\describe{
  \item{data_type}{test}
  \item{table_name}{Table identifier}
  \item{description}{Text}
  \item{organization}{Text}

  ...
}
}
\usage{
test_table_dict
}
\description{
A data frame with database schema information representative of DRIVES tables.
}
\keyword{datasets}
