% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unitacre_to_unithectare}
\alias{unitacre_to_unithectare}
\title{Convert any unit per acre area to the same unit per hectare area}
\usage{
unitacre_to_unithectare(unitacre)
}
\arguments{
\item{unitacre}{A value or numeric vectors in a per-acre unit, e.g., seeds/acre.}
}
\value{
A value or numeric  vector in the equivalent per-hectare unit, e.g., seeds/ha
}
\description{
Convert any unit per acre area to the same unit per hectare area
}
