% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_qc_functions.R
\name{check_dictionary}
\alias{check_dictionary}
\title{Check dictionary metadata}
\usage{
check_dictionary(
  table_name = "site_info",
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{table_name}{Table (collection) name.}

\item{mytoken}{API token, formatted as "Bearer apitoken"}

\item{...}{parameters passed to get_db_info and api_request}
}
\value{
A data frame with a set of conditions, outcome (TRUE = pass or FALSE = fails), and 
fields that fail the condition. In case of mismatched field names, there are separate 
columns for fields in the column_dictionary (dcols) and Directus table schema (tcols).
}
\description{
Check for discrepancies between collection metadata stored in Directus
 and in data dictionaries.
}
\examples{
# set mytoken in function or with set_default_token()
# Not run:  site_info_check <- check_dictionary(table_name = "site_info") 

}
