% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{check_dups}
\alias{check_dups}
\title{Subset a dataframe for rows that are duplicated across a set of columns.}
\usage{
check_dups(mydf, checkcols = NULL, dupvec = NULL)
}
\arguments{
\item{mydf}{A dataframe.}

\item{checkcols}{A vector of column names that you want to check for duplicates. Will also work with column indices. Alternative to dupvec.}

\item{dupvec}{A vector to use to check for duplicate values. Length(dupvec) = nrow(mydf). Alternative to column names.}
}
\value{
Returns a list containing the number of unique column combinations that have duplicate values (ndups) and a subsetted dataframe containing all duplicated values.
}
\description{
Subset a dataframe for rows that are duplicated across a set of columns.
}
\examples{
testdf <- data.frame("col1" = c("spam","eggs","spam","spam"),"col2" = c(1,2,1,1),"col3" = c("cat","fish","dog","cat"),"col4" = c(1,2,3,4))
check_dups(testdf, checkcols = c("col1","col2"))
check_dups(testdf, checkcols = c(1,2,3))
}
