% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivesR-package.R
\docType{package}
\name{drivesR-package}
\alias{drivesR}
\alias{drivesR-package}
\title{drivesR: DRIVES Database Workflow}
\description{
Contains various functions to access and harmonize data from the DRIVES database. Also contains functions used to prepare data, perform quality checks, and manage the database structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://drives-project.github.io/drivesR/}
}

}
\keyword{internal}
