% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivesR-package.R
\docType{package}
\name{drivesR-package}
\alias{drivesR}
\alias{drivesR-package}
\title{drivesR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
