% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_utility_functions.R
\name{generate_directus_pat}
\alias{generate_directus_pat}
\title{Create Directus Personal Access Token}
\usage{
generate_directus_pat(
  useremail = NULL,
  userpassword = NULL,
  myurl = getOption("drivesR.default.url")
)
}
\arguments{
\item{useremail}{Email address registered with data.drives-network.org.}

\item{userpassword}{Account password.}

\item{myurl}{Default url (data.drives-network.org)}
}
\value{
Returns a text object "Bearer {newPAT}" that can be used as 
the input for set_default_token().
}
\description{
This function generates a personal access token (PAT) for a user 
who has been registered on data.drives-network.org. The PAT is is
an alternative to an API key, meant to be used for single sessions
with the API. The PAT expires after about 10 minutes and can be 
regenerated as many times as needed by the user.
}
\details{
If using this in a script, the user email and password should be read
from an external file.
}
\examples{
# not run: source("directus_creds.R") ## script that reads email and password,
#not run: mypat <- generate_directus_pat(# objects created in "directus_creds.R"
#                                          useremail = directus_email, 
#                                          userpassword = directus_password)
# not run: set_default_token(mypat)
# not run: test_api_token(mypat) # to check if it works. 
}
