% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_db_table}
\alias{get_db_table}
\title{Fetch an entire table from Directus}
\usage{
get_db_table(
  table_name = "test_cat_info",
  myurl = "https://data.drives-network.org",
  mytoken = "Bearer {myAPItoken}"
)
}
\arguments{
\item{table_name}{The table identifier within Directus}

\item{myurl}{The database url. By default, "https://data.drives-network.org"}

\item{mytoken}{The user-specific API token, in the format "Bearer {insertAPItoken}", without curly brackets}
}
\value{
A data frame containing all rows and columns of the specified table
}
\description{
Fetch an entire table from Directus
}
\examples{
testdf <- get_db_table("test_cat_info")

}
