% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_db_table}
\alias{get_db_table}
\title{Fetch an entire table from Directus}
\usage{
get_db_table(
  table_name = "site_info",
  myurl = getOption("drivesR.default.url"),
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{table_name}{The table identifier within Directus}

\item{myurl}{The database url. By default, "https://data.drives-network.org"}

\item{mytoken}{The user-specific API token, in the format "Bearer {insertAPItoken}", without curly brackets.
This can be set with set_default_token.}
}
\value{
A data frame containing all rows and columns of the specified table. 
If public is set to TRUE, the data will exclude sites and years not approved 
for public access. Also, data from Canadian sites will be fetched from a 
Canadian-hosted repository (Borealis).
}
\description{
Fetch an entire table from Directus
}
\examples{
#not run: testdf <- get_db_table("site_info") 

}
