% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_db_table}
\alias{get_db_table}
\title{Fetch an entire table from Directus}
\usage{
get_db_table(
  table_name = "site_info",
  myurl = getOption("drivesR.default.url"),
  mytoken = getOption("drivesR.default.directustoken"),
  in_batches = FALSE,
  batchsize = NULL
)
}
\arguments{
\item{table_name}{The table identifier within Directus}

\item{myurl}{The database url. By default, "https://data.drives-network.org"}

\item{mytoken}{The user-specific API token, in the format "Bearer {insertAPItoken}", without curly brackets}

\item{in_batches}{When set to FALSE, all rows are imported with a single API call. When set to TRUE,
rows are imported in batches specified by batchsize. Batched import may be useful 
for large tables.}

\item{batchsize}{Sets the number of rows per batch-import when in_batches = TRUE.}
}
\value{
A data frame containing all rows and columns of the specified table.
}
\description{
Fetch an entire table from Directus
}
\examples{
#not run: testdf <- get_db_table("site_info") # at once
#not run: testdf2 <- get_db_table("site_info", in_batches = TRUE, batchsize = 4)

}
