% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{get_db_table}
\alias{get_db_table}
\title{Fetch an entire table from Directus}
\usage{
get_db_table(
  table_name = "site_info",
  public = getOption("drivesR.default.public"),
  myurl = getOption("drivesR.default.url"),
  mytoken = getOption("drivesR.default.directustoken"),
  public_tables = getOption("drivesR.default.tablevec"),
  borealis_repo_info = NULL,
  dataverse_api = NULL
)
}
\arguments{
\item{table_name}{The table identifier within Directus}

\item{public}{TRUE or FALSE indicating whether the table should be read from the 
public or internal portions of the DRIVES database. This can be set as a global option.}

\item{myurl}{The database url. By default, "https://data.drives-network.org"}

\item{mytoken}{The user-specific API token, in the format "Bearer {insertAPItoken}", without curly brackets}

\item{public_tables}{Vector of table names that receive the prefix 'public_' when public = TRUE.}

\item{borealis_repo_info}{A dataframe containing Borealis file identifiers for each table_name. 
If NULL, this information is imported from Directus.}

\item{dataverse_api}{API key for the Borealis repository. Included for pre-publication troubleshooting. 
Once tables are published, this can be left as NULL.}
}
\value{
A data frame containing all rows and columns of the specified table. 
If public is set to TRUE, the data will exclude sites and years not approved 
for public access. Also, data from Canadian sites will be fetched from a 
Canadian-hosted repository (Borealis).
}
\description{
Fetch an entire table from Directus
}
\examples{
#not run: testdf <- get_db_table("site_info") # at once
#not run: testdf2 <- get_db_table("site_info", in_batches = TRUE, batchsize = 4)

}
