% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{import_db_tables}
\alias{import_db_tables}
\title{Bulk download database tables}
\usage{
import_db_tables(
  tablevec = NULL,
  save_locally = FALSE,
  savedir = ".",
  savename = "drives_dblist",
  ...
)
}
\arguments{
\item{tablevec}{A vector of table names to download. If NULL, the function downloads all DRIVES tables
available to the user.}

\item{save_locally}{TRUE or FALSE indicating whether the downloaded tables should be 
saved to an R data object.}

\item{savedir}{Directory path for saving locally. Defaults to the working directory.}

\item{savename}{File name for saving locally, excluding .Rda. Defaults to "drives_dblist"}

\item{...}{Arguments to be passed to get_db_table and get_db_info.}
}
\value{
A list of DRIVES database tables.
}
\description{
Bulk download database tables
}
\examples{
# not run: db <- import_db_tables()
}
