% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{import_db_tables}
\alias{import_db_tables}
\title{Bulk download database tables}
\usage{
import_db_tables(
  tablevec = getOption("drivesR.default.tablevec"),
  save_locally = FALSE,
  import_from_local = FALSE,
  savedir = ".",
  savename = "drives_dblist",
  public = getOption("drivesR.default.public"),
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{tablevec}{A vector of table names to download. If NULL, the function downloads all DRIVES tables
available to the user.}

\item{save_locally}{TRUE or FALSE indicating whether the downloaded tables should be 
saved to an R data object.}

\item{import_from_local}{If TRUE, the function reads in an R data object with the path savedir and savename
instead of downloading the tables from directus.}

\item{savedir}{Directory path for saving locally. Defaults to the working directory.}

\item{savename}{File name for saving locally, excluding .Rda. Defaults to "drives_dblist"}

\item{public}{TRUE if data are to be downloaded from the publicly available part of the DRIVES 
database. FALSE otherwise.
Can be set directly or though options.}

\item{mytoken}{Directus API token, formatted as "Bearer apitoken". Can be set with set_default_token()}
}
\value{
A list of DRIVES database tables.
}
\description{
Bulk download database tables
}
\examples{
# not run: db <- import_db_tables()
}
