% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{import_dictionary_tables}
\alias{import_dictionary_tables}
\title{Import dictionary tables
These do not require an API key.}
\usage{
import_dictionary_tables()
}
\arguments{
\item{myurl}{URL for the Directus database. Set as a default option (data.drives-network.org)}
}
\value{
A list of three dictionary tables. One describing database tables, one 
describing columns within tables, and one describing categories within columns.
}
\description{
Import dictionary tables
These do not require an API key.
}
\examples{
# not run: dict <- import_dictionary_tables()
}
