% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_wrangling.R
\name{is.ISOdate}
\alias{is.ISOdate}
\title{Check if dates are in ISO format.
Returns TRUE if a vector of character string dates is in a
recognizable YYYY-MM-DD format. This works with - or / separators.
It Also works with single digit month or day.}
\usage{
is.ISOdate(d)
}
\arguments{
\item{d}{a vector of dates in character format}
}
\value{
a vector of TRUE and FALSE indicating whether each date is in ISO format. NAs are returns as NA.
}
\description{
Check if dates are in ISO format.
Returns TRUE if a vector of character string dates is in a
recognizable YYYY-MM-DD format. This works with - or / separators.
It Also works with single digit month or day.
}
\examples{
myd = c("2024-01-01","2022-01-01","6/23/22","24-01-01", NA,"jbs")
is.ISOdate(myd)
}
