% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pipeline.R
\name{list_treatments_by_management_practice}
\alias{list_treatments_by_management_practice}
\title{List treatment components by management practice.}
\usage{
list_treatments_by_management_practice(
  site_treatment_type_info = NULL,
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{site_treatment_type_info}{A data frame with the site_treatment_type_info table. If left as
NULL, this table is imported from directus.}

\item{mytoken}{Directus token, formatted as "Bearer myapitoken". Set by set_default_token().}
}
\value{
A named list of treatment types corresponding to a set of core management practices described for
all sites. Useful for identifying columns in harmonized dataframes with all treatment types.
}
\description{
Uses information in the site_treatment_type_info table
}
\examples{
#not run: management_practice_list <- list_treatments_by_management_practice()
# not run: harmonized_treatments <- harmonize_treatments()
# If I want to do something with N fertility treatments, for example.
#not run: nfert <- harmonized_treatments[,c("site_id","treatmentID2","year",management_practice_list$`N fertility`)]
}
