% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_build_functions.R
\name{post_rows_in_batches}
\alias{post_rows_in_batches}
\title{Post rows in batches}
\usage{
post_rows_in_batches(
  table_name = "crop_yields",
  batchsize = 1000,
  inputdf = NULL,
  mytoken
)
}
\arguments{
\item{table_name}{Table identifier in the database.}

\item{batchsize}{Number of rows to be added at a time. Suggested amount is 1000 rows.
Progress is printed to the console.}

\item{inputdf}{Data frame of rows to be added. This should have passed quality control checks.}

\item{mytoken}{Directus token. Can be set with set_default_token()}
}
\description{
To get around issues of posting limits.
}
