% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_fetch_functions.R
\name{query_table_by_pk}
\alias{query_table_by_pk}
\title{Fetch table contents by primary keys
Function to query contents of a database table with a vector of primary key values.
Useful for double-checking before running 'modify_rows' or 'delete_rows'.
It could be useful for querying information across tables.}
\usage{
query_table_by_pk(
  table_name = NULL,
  pkvec = NULL,
  pkfield = "uid",
  public = getOption("drivesR.default.public"),
  mytoken = getOption("drivesR.default.directustoken")
)
}
\arguments{
\item{table_name}{The table identifier in the DRIVES database. If public = TRUE, the table name is automatically 
modified to query the public version of the database table.}

\item{pkvec}{A vector of primary key values.}

\item{pkfield}{The name of the column name that holds the table's primary key. 
For most tables, this is 'uid'.}

\item{public}{If TRUE, the function queries publicly available data tables. 
Since this function is mostly for internal use, the default is FALSE.}

\item{mytoken}{Directus token, formatted as "Bearer APITOKEN". 
Can be set with set_default_token.}

\item{myurl}{Directus URL. Set at package loading ("https://data.drives-network.org").}
}
\description{
Fetch table contents by primary keys
Function to query contents of a database table with a vector of primary key values.
Useful for double-checking before running 'modify_rows' or 'delete_rows'.
It could be useful for querying information across tables.
}
\examples{
#not run: qtable <- query_table_by_pk("harvest_dates",pkvec = 1:100, pkfield = "uid)
}
