% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directus_utility_functions.R
\name{set_default_token}
\alias{set_default_token}
\title{Set default API token}
\usage{
set_default_token(usertoken)
}
\arguments{
\item{mytoken}{User-specific API token formated as "Bearer {myAPI token"}}
}
\value{
Send
Resets default api token for relevant functions in drivesR by sending commands to the R console.
For now, it just specifies relevant functions in separate lines of code. 
There's probably a fancier way to do this with lists or whatever.
}
\description{
Set a user-specific API token as the default for functions making API requests. 
Only works in R studio.
}
\examples{
directus_token = "Bearer abunchofnumbersandletters"
## This token should be read from a script or text file that is not synced to github

# see what defaults are before and after running
formals(api_request)$mytoken
# Not run: set_default_token(directus_token)
formals(api_request)$mytoken

}
